requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
psutil>=5.8.0
requests>=2.26.0
python-dotenv>=0.19.0
pydantic>=1.8.2
aiohttp>=3.8.1
websockets>=10.0
python-multipart>=0.0.5
PyYAML>=6.0.0
rich>=10.0.0
typing-extensions>=4.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-timeout>=2.1.0
black>=23.0.0
isort>=5.0.0
mypy>=1.0.0
flake8>=6.0.0
tox>=4.0.0

# Documentation dependencies
mkdocs>=1.4.0
mkdocs-material>=9.0.0
mkdocstrings>=0.20.0
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0
sphinx-autodoc-typehints>=1.12.0
sphinx-copybutton>=0.5.0
myst-parser>=0.15.0 